Removed weird construct in OpenTimestamps.verify(). Removed a ton of unnecessary verbose "this." in Timestamp. Replace a few loop additions of elements with addAll(). Improved string concatenation performance by linking calls to StringBuilder::append(). Fixed bug in Timestamp.equals(). Replaced all use of Utils.toBytes() with org.bitcoinj.core.Utils.toBytes(). Upgraded to JUnit 4.12, which gives me assertArrayEquals(), which I now use.

+++ src/test/java/com/eternitywall/TestOpenTimestamps.java
@@ -160,7 +160,7 @@ public class TestOpenTimestamps{
                 HashMap<VerifyResult.Chains, VerifyResult> results = OpenTimestamps.verify(detachedOts, detached);
                 assertTrue(results.size()>0);
                 assertTrue(results.containsKey(VerifyResult.Chains.BITCOIN));
-                assertEquals(1432827678L, results.get(VerifyResult.Chains.BITCOIN).timestamp);
+                assertEquals(1432827678L, results.get(VerifyResult.Chains.BITCOIN).timestamp.longValue());
             } catch (Exception e){
                 fail(e.toString());
             }
